+++ src/griptape_nodes/servers/static.py
